Adds check ownership functionality to the update and delete methods for a link

+++ project/app/schema.py
@@ -1,3 +1,4 @@
+from django.contrib import auth
 import graphene
 from graphene_django import DjangoObjectType
 from users.schema import UserType
@@ -52,27 +53,42 @@ class UpdateLink(graphene.Mutation):
     description = graphene.String()
     owner = graphene.Field(UserType)
 
+
     class Arguments:
         id = graphene.ID()
         url = graphene.String()
         description = graphene.String()
-
+        
     def mutate(self, info, url, description, id):
         user = info.context.user
-        owned_links = Link.objects.filter(owner=info.context.user).values_list("owner_id", flat=True).first()
-        print(owned_links, user.id)
-        if user.is_authenticated:
-            if owned_links == user.id:
-                link = Link(id=id, url=url, description=description)
+        # This provides a query set of all of the users owned links
+        owner_links = Link.objects.filter(owner=info.context.user).values()
+        # print(owner_links)
+        # iterate through the list to check each value
+        for owner in owner_links:
+            # first we extract the owner as a list
+            owner = list(owner.values())
+            # the owner is the last value in the list
+            owned_id = (owner[0])
+
+            print([owned_id, int(id)])
+            # create a list to store the values of the id in the list of owned ids and the id passed in
+            is_owner = [owned_id, int(id)]
+            # compare the id of the owned links to the id passed in and if the user is logged in
+            if (is_owner[0] == is_owner[1]) and user.is_authenticated:
+                # print('authorized')
+                # mutate the link based on the arguments below
+                link = Link(id=id, url=url, description=description, owner=user)
+                # save the link
                 link.save()
+                # then return it
                 return UpdateLink(
                     id=link.id,
                     url=link.url,
                     description=link.description,
                     owner=link.owner
-                    )
-            raise Exception('Not authorized to update this link')
-        raise Exception('Please Log in')
+                )
+        raise Exception('Not authorized to update this link. Please sign in or try a different link.')
 class DeleteLink(graphene.Mutation):
     id = graphene.Int()
 
@@ -80,12 +96,30 @@ class DeleteLink(graphene.Mutation):
         id = graphene.Int()
 
     def mutate(self, info, id):
-        link = Link(id=id)
-        link.delete()
+        user = info.context.user
+        # This provides a query set of all of the users owned links
+        owner_links = Link.objects.filter(owner=info.context.user).values()
+        # print(owner_links)
+        # iterate through the list to check each value
+        for owner in owner_links:
+            # first we extract the owner as a list
+            owner = list(owner.values())
+            # the owner is the last value in the list
+            owned_id = (owner[0])
 
-        return DeleteLink(
-            id=link.id      
-        )
+            print([owned_id, int(id)])
+            # create a list to store the values of the id in the list of owned ids and the id passed in
+            is_owner = [owned_id, int(id)]
+            # compare the id of the owned links to the id passed in and if the user is logged in
+            if (is_owner[0] == is_owner[1]) and user.is_authenticated:
+                link = Link(id=id)
+                link.delete()
+
+                return DeleteLink(
+                    id=link.id      
+                )
+        raise Exception('Not authorized to delete this link. Please sign in or try a different link.')
+
 
 class Mutation(graphene.ObjectType):
     create_link = CreateLink.Field()
